Finish spgemm perf test refactor
- Add check_arg_double(...) to perf test utils
- Use perf test utils for arg parsing, selecting backend
- Get rid of the ability to have distinct fast/slow memory spaces
- Merge the three files that made up the perf test into just
  KokkosSparse_spgemm.cpp
- Change some types in KokkosKernels::Experiment::Parameters to be safer
  - on/off values should be bool, not int
  - strings should be std::string, not char*
