+++ perf_test/sparse/KokkosSparse_spgemm_jacobi.cpp
@@ -219,12 +219,12 @@ int main(int argc, char** argv) {
   if (parse_inputs(params, argc, argv)) {
     return 1;
   }
-  if (params.a_mtx_bin_file == NULL) {
+  if (params.a_mtx_bin_file == "") {
     std::cerr << "Provide a and b matrix files" << std::endl;
     print_options();
     return 0;
   }
-  if (params.b_mtx_bin_file == NULL) {
+  if (params.b_mtx_bin_file == "") {
     std::cout << "B is not provided. Multiplying AxA." << std::endl;
   }
